+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Fixed an issue where the **create-id-set** command did not identify unified integrations correctly.
 * Fixed an issue where the `CommonTypes` pack was not identified as a dependency for all feed integrations.
 * Added support for running SDK commands in private repositories.
-* Added a validation to ensure correct image file names.
+* Added a validation to ensure correct image and description file names.
 
 # 1.3.5
 * Added a validation that layoutscontainer's id and name are matching. Updated the format of layoutcontainer to include update_id too.